requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Security
PyJWT>=2.8.0
cryptography>=41.0.0
passlib[bcrypt]>=1.7.4

# Email Service
resend>=0.7.0

# Utils
python-multipart>=0.0.6
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
